[PP][SUPP-64] cleanup
